Add ability to plug in user-defined extensions.
`Extension` is turned into an enum, which can be extended to
include all RFC-specified extensions in addition to user-defined ones.

However, this patch only adds a way to extend Extensions, and a
Unknown variant.

Note: the extra `serde(bound)` is necessary to stop `serde` from
deriving `Default` for `E`. We don't want to constrain the user
unnecessarily.