User-defined extension support (#12)
* Add ability to plug in user-defined extensions.

`Extension` is turned into an enum, which can be extended to
include all RFC-specified extensions in addition to user-defined ones.

However, this patch only adds a way to extend Extensions, and a
Unknown variant.

Note: the extra `serde(bound)` is necessary to stop `serde` from
deriving `Default` for `E`. We don't want to constrain the user
unnecessarily.

* Default user-defined extensions to unit type.

This is so that we stay somewhat backwards compatible and we make the
API simpler to use in case the caller doesn't have user-defined extensions.