Propose impl from for editable fields (#53)
It would be nice to implement From<Inner> for EditableField<EditableFieldInner> as a shortcut when you don't have any metadata to add to the editable field.

This drastically simplifies our internal usage and I expect we're not the only provider which do not track ID or offer the label for many fields.